feature 4: user can update a note
