+++ node-notes-cli/data.json
@@ -1,9 +1,10 @@
 {
-  "nextId": 7,
+  "nextId": 8,
   "notes": {
-    "1": "The event loop is how a JavaScript runtime pushes asynchronous callbacks onto the stack once the stack is cleared.",
+    "1": "an update",
     "2": "Prototypal inheritance is how JavaScript objects delegate behavior.",
     "3": "In JavaScript, the value of `this` is determined when a function is called; not when it is defined.",
-    "5": "this is a note"
+    "5": "this is a note",
+    "7": "The JSON.stringify() method converts a JavaScript object or value to a JSON string"
   }
 }
